Migrate GPU torch models (RealMLP, xRFM) to per-model folder layout
Same TabSTAR recipe: `hpo.py` becomes the canonical home for each model's
search-space generator (including the `generate_configs_*` / `generate_single_config_*`
helpers that the previous `generate.py` defined inline); `info.py` carries
the MethodMetadata + ModelInfo bundle. `generate.py` shims preserve the
legacy `models/utils.py:name_to_import_map` dispatch.

The `realmlp_gpu_metadata` and `xrfm_metadata` entries in
`_tabarena_method_metadata_2025_09_03.py` are now sourced from each model's
`info.py` (legacy names kept as aliases).
