+++ tabarena/tabarena/models/realmlp/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.realmlp.hpo import gen_realmlp
+from tabarena.models.realmlp.info import realmlp_info, realmlp_method_metadata
+
+__all__ = ["gen_realmlp", "realmlp_info", "realmlp_method_metadata"]
