+++ tabarena/tabarena/models/xrfm/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.xrfm.hpo import gen_xrfm
+from tabarena.models.xrfm.info import xrfm_info, xrfm_method_metadata
+
+__all__ = ["gen_xrfm", "xrfm_info", "xrfm_method_metadata"]
